Anchor Center changes WIP

+++ app/src/main/java/com/rgb/matrix/GameOver.java
@@ -11,7 +11,7 @@ import org.andengine.entity.text.Text;
 import org.andengine.input.touch.TouchEvent;
 import org.andengine.opengl.font.Font;
 import org.andengine.opengl.vbo.VertexBufferObjectManager;
-import org.andengine.util.color.Color;
+import org.andengine.util.adt.color.Color;
 import org.andengine.util.modifier.IModifier;
 
 /**
@@ -71,7 +71,7 @@ public class GameOver extends Entity {
         return GAMEOVER_DIALOG_WIDTH;
     }
 
-    public int getHeight() {
+    public float getHeight() {
         return GAMEOVER_DIALOG_HEIGHT;
     }
